delete should work now
