+++ ycms/cms/templates/ward/ward_management.html
@@ -132,17 +132,19 @@
                                 <td>{{ ward.total_beds }}</td>
                                 <td>{{ ward.available_beds }}</td>
                                 <td>{{ ward.occupied_beds }}</td>
-
                                 <td align="left">
-                                    <button type="submit" class="btn w-full md:w-auto self-center">{% translate "Edit" %}</button>
-                                </td>
-                                <td align="left">
-
-                                        {% csrf_token %}
-                                        <button type="submit" class="text-white bg-red-500 rounded px-4 py-2.5">
+                                    <button class="btn w-full md:w-auto self-center">{% translate "Edit" %}</button>
+                                    <button class="grow text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 rounded px-5 py-2.5 dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-900"
+                                            form="delete-ward-{{ ward.id }}"
+                                            onclick="return confirm('{% translate "Are you sure you want to delete this ward?" %}')">
+                                        <form hidden
+                                              id="delete-ward-{{ ward.id }}"
+                                              method="post"
+                                              action="{% url "cms:protected:delete_ward" ward.id %}">
+                                            {% csrf_token %}
+                                        </form>
                                         {% translate "Delete" %}
-                                        </button>
-
+                                    </button>
                                 </td>
 
                             </tr>
