+++ ycms/cms/urls/protected.py
@@ -14,7 +14,6 @@ from ..views import (
 )
 from ..views.floor.floor_view import FloorView  # Importiere die FloorView hier
 from ..views.utility.autocomplete import autocomplete_icd10, autocomplete_patient
-from ..views.ward.ward_edit import WardDeleteView
 
 urlpatterns = [
     path("", index.UserBasedRedirectView.as_view(), name="index"),
@@ -113,6 +112,9 @@ urlpatterns = [
                 path("<int:pk>/", ward.WardView.as_view(), name="ward_detail"),
                 path(
                     "manage/", ward.WardManagementView.as_view(), name="ward_management"
+                ),
+                path(
+                    "delete/<int:pk>", ward.WardDeleteView.as_view(), name="delete_ward"
                 ),
             ]
         ),
