+++ ycms/cms/views/ward/__init__.py
@@ -1,6 +1,5 @@
 """
 This package contains views related to the ward
 """
-from .ward_edit import WardDeleteView
-from .ward_management_view import WardManagementView
+from .ward_management_view import WardDeleteView, WardManagementView
 from .ward_view import WardView
